pyproject.toml
[project]
name = "stock-agent-orchestration"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12.0"
dependencies = [
    "a2a-sdk>=0.3.9",
    "bs4>=0.0.2",
    "feedparser>=6.0.12",
    "google-adk>=1.16.0",
    "html2text>=2025.4.15",
    "httpx>=0.28.1",
    "mcp>=1.17.0",
    "nltk>=3.9.2",
    "pandas>=2.3.3",
    "rich>=14.2.0",
    "torch>=2.9.0",
    "transformers>=4.57.1",
    "uvicorn>=0.37.0",
    "yfinance>=0.2.66",
]
